requirements.txt
pandas
numba
Quaternion
scipy
numpy
Pillow
matplotlib
seaborn
progress
opencv
